Extract errors and add exception handler for ModelServer
