+++ python/kserve/kserve/errors.py
@@ -1,0 +1,74 @@
+# Copyright 2022 The KServe Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from http import HTTPStatus
+
+from fastapi.responses import JSONResponse
+
+
+class ModelMissingError(Exception):
+    def __init__(self, path):
+        self.path = path
+
+    def __str__(self):
+        return self.path
+
+
+class InferenceError(RuntimeError):
+    def __init__(self, reason):
+        self.reason = reason
+
+    def __str__(self):
+        return self.reason
+
+
+class InvalidInput(ValueError):
+    """
+    Exception class indicating invalid input arguments.
+    HTTP Servers should return HTTP_400 (Bad Request).
+    """
+
+    def __init__(self, reason):
+        self.reason = reason
+
+    def __str__(self):
+        return self.reason
+
+
+class ModelNotFound(Exception):
+    """
+    Exception class indicating requested model does not exist.
+    HTTP Servers should return HTTP_404 (Not Found).
+    """
+
+    def __init__(self, model_name=None):
+        self.reason = f"Model '{model_name}' not found."
+
+    def __str__(self):
+        return self.reason
+
+
+async def exception_handler(_, exc):
+    return JSONResponse(status_code=HTTPStatus.INTERNAL_SERVER_ERROR, content={"detail": str(exc)})
+
+
+async def invalid_input_handler(_, exc):
+    return JSONResponse(status_code=HTTPStatus.BAD_REQUEST, content={"detail": str(exc)})
+
+
+async def inference_error_handler(_, exc):
+    return JSONResponse(status_code=HTTPStatus.INTERNAL_SERVER_ERROR, content={"detail": str(exc)})
+
+
+async def model_not_found_handler(_, exc):
+    return JSONResponse(status_code=HTTPStatus.NOT_FOUND, content={"detail": str(exc)})
