+++ python/lgbserver/lgbserver/__main__.py
@@ -16,7 +16,7 @@ import argparse
 import logging
 import kserve
 
-from kserve.model import ModelMissingError
+from kserve.errors import ModelMissingError
 from lgbserver.lightgbm_model_repository import LightGBMModelRepository
 from lgbserver.model import LightGBMModel
